Add food diary working

+++ DietPlanner.Contracts/Models/AddDietEntry.cs
@@ -3,7 +3,7 @@ public class AddDietEntry
 {
     public string UserID { get; set; }
     public string FoodName { get; set; }
-    public int MealTypeID { get; set; }
+    public long MealTypeID { get; set; }
     public int Calories { get; set; }
     public DateTime Date { get; set; }
 }
